Show --help when CLI is called with no arguments
Currently there is an issue (documented in #69) where a cryptic error
is displayed when the CLI is called with no arguments. While the use
of `--help` is documented in the README, it would be more user-friendly
for the CLI to default to showing the usage message if no argument is
provided at invocation.

Fixes #69