Correct interpretation of `negative` metadata
In October of 2016, the frontmatter format of Test262 tests was updated
to include more information about expected errors [1]. Prior to that
change, the `negative` attribute was defined as a regular expression
that matched the expected error's `name` attribute. Following that
change, the `negative` attribute is defined as a dictionary containing
`phase` and `type` attributes. The `type` attribute is a string value
describing the expected error's `name` attribute.

Update the validation logic to interpret the meta data according to this
new format.

[1] https://github.com/tc39/test262/pull/778