Ensure version compatibility
This is a breaking change; details are provided below.

Because the instructions for correctly interpreting Test262 tests
occasionally change, `test262-harness` may not be capable of correctly
interpreting the input files.

Previously, `test262-harness` would accept test files without regard for
the corresponding version of Test262. This made it possible for users to
unknowingly use incompatible versions and thereby receive invalid
results.

Not all changes in the interpretation guidelines effect all users. For
instance, a user interested in testing a "polyfill" for
`Array.prototype.includes` would not be effected by alterations to
interpretation guidelines for Atomics. Supporting use cases like this
involves allowing users to *intentionally* use incompatible versions of
Test262.

Inspect the version of Test262 under test and ensure that it is within a
known-compatible range. Implement a new option named `--acceptVersion`
to allow users to explicitly opt out of this behavior.

Note on backwards-compatibility:

Previously, arbitrary files could be tested as long as an "includes"
directory was specified. This change is not backwards-compatible because
it requires one of the following:

- input files are located within a Test262 directory
- a Test262 directory is specified via the `--test262Dir` command-line
  option