Decrease timeout during testing
The project's test suite includes test material that intentionally cause
timeouts. This material is executed multiple times under different
configurations, so the default value of 5000 milliseconds drastically
increases the time required to execute the entire test suite.

Reduce the "timeout" value used when testing the project so that tests
which are expected to cause a timeout do so more expediently.